Stop timer to avoid suspected race condition in tests
* We suspect that there is a race in the tests when
WaitForWatcherAndIncrement is called. If the timer isn't stopped, it may
still count as a 'watcher' after it has written something to its
channel. To get around this, we're experimenting with stopping the timer
as soon as timer's channel receives something.
* We haven't been able to reproduce the race ourselves, but we hope that
this will stop the particular build breaks in CI that we've been seeing
around this code.

Signed-off-by: Jen Spinney <jen.spinney@hpe.com>