Improve readability and flakiness of the watcher test suite
this commit gets rid of the ready channel that is used to force bbs events to
be cached during a sync event. the new approach is to send the event inside one
of the bbs client stubs. the commit also reorganize contexts; most of the
contexts and tests didn't require the event caching functionality anyway

also fixed a few eventually that were passing the return value of the functions
instead of the function itself. this became obvious after getting rid of the
ready channel since it used to serve as a synchronization point; in other words
the eventually used to be redundant due to the use of the ready channel

[#142850187]