add a test to demonstrate the bug on warp-drive
the bug is caused by the short-circuit logic for the internal routes. the short
circuit logic is triggered if there are multiple port mapping defined on the
LRP (e.g. sshd port and app port). Since internal routes don't have an
associated port the loop will initially remove the internal route and the next
iteration will realize that the internal route has been removed and
short-circuit the function causing all but the first port mapping to be
unregistered. This sometimes go unnoticed if the first port mapping is the app
port followed by the sshd port, since the sshd port don't have any routes
associated with it.

[#153082094]