ignore EndpointUpdatedAtNs in the integration test
this is problematic to test since a DesiredLRP create event followed by
ActualLRP update event may be handled as part of a sync loop. in this case
EndpointUpdatedAtNs is not being set. Alternatively, RE could handle those two
events separately causing the EndpointUpdatedAtNs to be set to the Since
field.

[#153070295]