Use flattened ActualLRP BBS endpoints instead of ActualLRPGroup endpoints
Everywhere in the code, we replaced all references to ActualLRPGroups
with flattened ActualLRPs. We also replaced all references to ActualLRP
events that used ActualLRPGroups to ActualLRPInstance events.

This commit introduces slightly different behavior for the
route-emitter: we no longer collapse multiple ActualLRPs with the same
process GUID and index into a single routing endpoint (as we did by
resolving the ActualLRPGroup).  Instead, we now handle each ActualLRP on
a per-instance basis.  This could potentially mean that we could get
extra endpoints in the routing table with the same process GUID and
index, but different presences.

However, with a route emitter that filters by cell ID, we do not expect
more than one ActualLRP with a given process GUID and index because a
different presence implies that the LRP must be on a different cell.
In the old behavior with ActualLRPGroups, we wouldn't have seen any more
than one ActualLRP in the group due to the filtering, so the
ActualLRPGroup was effectively already an ActualLRP instance anyways.

By relying on ActualLRPInstance events, the number of events we'll
handle will differ from the number of events we handled when we used
ActualLRPGroup events. Instead of looking at the ActualLRP group and
implicitly adding or removing routes based on the changes we observed in
the group, we will now explicitly add or remove routes based on the
event.

This change should simplify the logic in the route-emitter, since it is
now the BBS's responsibility for adding or cleaning up individual
routing endpoints (through a 1-1 correspondence between flattened
ActualLRP and routing endpoint).

[#159596156]

Co-authored-by: Edwin Xie <exie@pivotal.io>