Fix failing tests
Context: https://ci.funtime.lol/teams/wg-arp-diego/pipelines/diego-release/jobs/unit-and-integration-tests/builds/118

After Adding cloudfoundry/route-emitter#34 it takes up-to 30Sec to
connect with NATS. This was causing an issue for the tests when it
wasn't wait long enough to have the service up and running.

Additionally, killing nats server in the test to run it with TLS was
failing the retry logic. Instead, we are always running NATS with TLS in
the suite.